rag-search/deploy/requirements.txt
beautifulsoup4
chromadb==1.0.6
googlesearch-python==1.3.0
groq==0.19.0
python-dotenv==1.1.0
requests==2.32.3
gradio==5.25.0
google_search_results==2.4.2
